+++ storage/node/node-api/transport.js
@@ -14,7 +14,7 @@ const authorize = require("../authorize");
 const Roles = require("../roles");
 const config = require('../config');
 const logger = require('../../utils/logger');
-const { Storage } = require('@dictadata/storage-junctions');
+const { Storage } = require('@dictadata/storage-tracts');
 const { StorageResults, StorageError } = require('@dictadata/storage-junctions/types');
 const stream = require('stream').promises;