+++ storage/node/tracts.js
@@ -6,7 +6,7 @@
  */
 "use strict";
 
-const { Codex } = require("@dictadata/storage-junctions");
+const { Storage } = require("@dictadata/storage-tracts");
 const logger = require("../utils/logger");
 const fs = require("fs");
 
@@ -15,16 +15,15 @@ const fs = require("fs");
  */
 exports.startup = async (config) => {
   logger.info("tracts startup");
-  logger.verbose("tracts SMT: " + JSON.stringify(config.codex.tracts.smt));
+  logger.verbose("tracts SMT: " + JSON.stringify(config.tracts.smt));
 
   var exitCode = 0;
 
   try {
-    if (config.codex.tracts) {
+    if (config.tracts) {
       // create tracts
-      let tracts = Codex.use("tract", config.codex.tracts.smt, config.codex.tracts.options);
-      await tracts.activate();
-      await addTracts(config.codex.tracts.tracts_cache);
+      await Storage.tracts.activate(config.tracts.smt, config.tracts.options);
+      await addTracts(config.tracts.tracts_cache);
     }
   }
   catch (err) {
@@ -35,7 +34,7 @@ exports.startup = async (config) => {
   return exitCode;
 };
 
-// add tracts entries from config.codex.tracts.tracts_cache (local cache only)
+// add tracts entries from config.tracts.tracts_cache (local cache only)
 async function addTracts(tracts_cache) {
   if (Array.isArray(tracts_cache)) {
     for (let entry of tracts_cache) {