+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG for @dictadata/storage-node
 
+- version 0.9.73
+  - refactor: reorg storage-junctions and storage-tracks
+  - updated: @dictadata/storage-junctions@0.9.98
+  - updated: @dictadata/storage-tracts@0.9.71
 - version 0.9.72
   - updated: @dictadata/storage-junctions@0.9.97
   - refactor: change Storage to be part of library interface